# perf: avoid graph break for SiLUT when inferring